Adding nick name voting
this was quite a learning experience around how JS and ko
work.  First off, I had the idea to allow clicking on the nick
names, expanding the quiz concept, to increment a counter for the
nick names as well, like voting.  However, it proved much more dificult
than the example I was using from KO's website.  In the foreach
functionality of the observeable array, I tried a number of things,
changing this to self in the main code, so I could use this for the
object getting passed back from the click in the foreach loop,
using other kinds of functions to increment the clicks, ultimately
leanrning that observeableArrays don't update only if their
properties update. So, I then used the replace function, but also
had to allow the replace function to see the orignal, and make a copy of
the object to increase the clicks.  Ultimlatly, that's where I landed.