create a way to see the current searcn and ...
remove the auto-apply filter when adding a filter item.
With the current setup, it made removing and apply the filter hard,
so it was inconsistant to be able to add the filter without applying.

Also updated the list of parks, and made the links to NYT open new
targets instead.